separate the tests of the echo-get example
